refactor S3 reading logic and move files to tools/
1. Move s3 utils torch/testing/_internal/* to tools/
2. Create tools/test/test_stats to test logics in tools
3. make run_test.py import discovery safe-guard
