+++ mypy-strict.ini
@@ -37,10 +37,11 @@ strict_equality = True
 files = tools/codegen/gen.py,
     tools/autograd/*.py,
     tools/pyi/*.py,
+    tools/stats_utils/*.py,
     tools/test_history.py,
+    tools/print_test_stats.py,
     torch/testing/_internal/framework_utils.py,
     torch/testing/_internal/mypy_wrapper.py,
-    torch/testing/_internal/print_test_stats.py,
     torch/utils/benchmark/utils/common.py,
     torch/utils/benchmark/utils/timer.py,
     torch/utils/benchmark/utils/valgrind_wrapper/*.py,
