+++ mypy.ini
@@ -35,7 +35,8 @@ files =
     test/test_type_info.py,
     test/test_utils.py,
     tools/generate_torch_version.py,
-    tools/clang_format_utils.py
+    tools/clang_format_utils.py,
+    tools/stats_utils/*.py
 
 
 # Minimum version supported - variable annotations were introduced
